@@ -10,7 +10,7 @@ authors = [
 ]
 description = "AI-enhanced Differentiable Ray Tracer for Irradiation Prediction in Solar Tower Digital Twins"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: MIT License",
@@ -26,7 +26,6 @@ dependencies = [
 [project.optional-dependencies]
 dev = [
     "pre-commit",
-    "black",
     "ruff",
     "pytest",
     "pytest-cov",
@@ -36,7 +35,7 @@ dev = [
     "sphinxemoji",
 ]
 mpi = ["mpi4py",]
-tutorials = ['jupyter', 'matplotlib']
+tutorials = ["matplotlib"]
 
 [project.urls]
 Homepage = "https://github.com/ARTIST-Association/ARTIST"
@@ -77,8 +76,8 @@ exclude = [
 line-length = 88
 indent-width = 4
 
-# Assume Python 3.8.
-target-version = "py38"
+# Assume Python 3.11.
+target-version = "py311"
 
 [tool.ruff.lint]
 # Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
