@@ -39,7 +39,7 @@ html_theme = "sphinx_rtd_theme"
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ["_static/"]
 html_css_files = ["custom.css"]
-html_logo = f"../logos/artist_logo.svg"
+html_logo = "../logos/artist_logo.svg"
 
 html_theme_options = {
     "logo_only": True,
